Update patch set 1
Patch Set 1:

So, the reason it was the way it was was so that it would also work with installs into a prefix.  This is how I normally run, so it's important to me.  I'll give this a try to see if it still works with prefixes.  (was)

Patch-set: 1
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>