Clean up UTF-8 checking in preparation for config option
UTF-8 validation of path components will become optional with a
subsequent patch.

Symlink values MUST NOT be validated.

Compound tags should only be checked for valid UTF-8, there is no
reason to not allow "." and ".." strings or '/' characters.

At the momenent we don't do ANY UTF-8 validation for the strings
that RFC 7530 says MUST be valid UTF-8.

Change-Id: Ie6d1fbb89148168715b95b3d907e37f852a6064e
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>