Update patch set 2
Patch Set 2:

So this change just 1) adds the option 2) defaulting to false

So that makes it harder to see (need to actually read code) whether currently this validation
is done only for NFSv4 (which I think should certainly be the case).

That said, I am sympathetic to the view that validating utf-8 strings is intrusive, and creates issues for at least legacy clients with certain kinds of data.  I'd be inclined to at least allow the option to disable this.

Patch-set: 2
Reviewer: Gerrit User 1004075 <1004075@d5d70762-12d0-45a1-890d-524b12d3f735>