Update patch set 4
Patch Set 4:

I believe this is the correct oa_flavor.  It is set if we successfully got through RPCSEC_GSS_INIT.  Maybe it should be set sooner?  But we will have gotten an error back in that case anyway.

no_dispatch is only set for RPCSEC_GSS, but it's initialized to false, so it's safe to unconditionally check.

r_u.req.svc.rq_verf.oa_flavor hasn't been in ntirpc since 1.4

Patch-set: 4