State - Allow last ref to be dropped in _state_del_locked
If an entry was killed (for example for being stale) then the last ref
could be held by the state.  This would lead to a use-after-free in
_state_del_locked() when allowing the cleanup.

However, we've completely unlinked and unlocked before we drop that last
ref, so it's safe to do the cleanup during that unref, since it won't
recurse for this state.

Change-Id: I3454223536cd68e700a6b23c7476c35b3dcab804
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>