Convert POSIX ACL to extented attribute
Add new generic interfaces for converting POSIX ACL to extented
attribute(and vice versa).

We are using the CephFS as storage backend. Althougth it's a
POSIX-compliant file system, it has no getacl and setacl interface
available. The getxattr and setxattr could be used to get or set ACL in
the CephFS.

Fortunately, the API fsal_acl_2_posix_acl and posix_acl_2_fsal_acl could
be used to covert NFSv4 acl to posix acl(and vice versa).

So we implement the API xattr_2_posix_acl and posix_acl_2_xattr in order
to support POSIX ACL when the CephFS is used as storage backend in
nfs-ganesha.

Change-Id: I7428cb9298b0041a524a1db267b69f2768202def
Signed-off-by: Shangzhong Zhu <zhu.shangzhong@zte.com.cn>
Signed-off-by: Luo Muyao <luo.muyao@zte.com.cn>