Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-2
> 
> Why not actually store an NFS v4 ACL in Ceph instead of an unreliably converted POSIX ACL?
> 
> However, either way, there's a problem with doing this.
> 
> Ganesha relies on the underlying filesystem to do permission checks for directory mutating operations (create, link, unlink, rename) because it's impossible to atomically check permission and perform an operation. The biggest hole is on a rename.
> 
> If you have a rename "a" to "b", the permission checks out ok to unlink "b" (or it doesn't currently exist), but before Ganesha does the rename, another process causes a new "b" to be created that the process doing the rename through Ganesha doesn't have permission to unlink, Ganesha will (using root privileges) replace that new "b" with "a".

Frank, thanks for the review. 
Only POSIX ACL is supported in the CephFS. So we have to convert NFSv4 ACL to POSIX ACL first. As you said, Ganesha relies on the underlying filesystem to do permission checks. Although the NFSv4 ACL can be coverted to extended attribute and stored to CephFS, the CephFS can only check the POSIX ACL permission.

Patch-set: 9