Update patch set 12
Patch Set 12:

(1 comment)

So, the reason I'm wondering about the application requirements is that CephFS doesn't support ACLs, so these won't be properly enforced.  Ganesha has a best-effort ACL enforcement, but it depends on the underlying filesystem to do proper ACL enforcement.  Gluster has this, as does GPFS, but CephFS does not.  This means there will be a window where an ACL can change, but Ganesha will still be enforcing the old ACL.  That is, ACL enforcement is not atomic.

This means you cannot depend on these ACLs for security uses.  They can only be used for convenience.   It also means we should probably default this option to OFF, but I could be convinced otherwise.

Patch-set: 12