Update patch set 1
Patch Set 1: Code-Review-1

No, the code is correct.  mdcache_avl_lookup_ck() takes a ref on the chunk containing the returned dirent (which is "next" in this case).  We don't want that ref, so we need to drop it.  next->chunk is the correct chunk to drop; it's not the next chunk (which would be chunk->next) but the chunk containing "next" (which is a dirent).  Definitely confusingly named variables, but correct code.

Patch-set: 1
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1