Don't convert posix acl to fsal acl when no access acl
When e_acl is NULL, the posix_acl_2_fsal_acl will return -1.
The EFAULT will be returned to the caller.
In fact, no access acl is a normal case for the file or directory.
So we only convert posix acl to fsal acl when the access acl is not NULL.

Signed-off-by: Zhu Shangzhong <zhu.shangzhong@zte.com.cn>
Change-Id: Icdc7756a5128192fb026c7f34dbf775999460770