MOUNT3 and LOOKUP3 partly working hardcoded inputs
The first thing the NFS Proxy has to do is run MOUNT. Normally this
involves a trip through the portmapper to guess the port (2050 by
default), so we just skip all that and issue the RPCs we need
directly.

To verify that the result is even remotely correct though, the next
thing Ganesha wants is LOOKUP. So, even though this change doesn't
fill in the result from the LOOKUP, nor even pass through credentials
properly, it does correctly send the root filehandle returned from
MOUNT (verified via watching tshark).

Change-Id: I0dd3f2129c1dae0079aab7982734a830a2430c39
Signed-off-by: Solomon Boulos <boulos@google.com>