Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (1 comment)

Thank you very much, but I think my understanding should be correct.

First, LogFatal() - a very serious problem uses this function to log. malloc(0) return NULL，which is actually wrong.

Then, If I see ”Ganesha assumes malloc(0) returns a non-NULL pointer” in the log file. The message I got was that the Ganesha process thought that a non-NULL pointer was returned through malloc(0), which was a serious error [by Logfatal()]. On the contrary, return null should be a error.

Patch-set: 2