Supporting protocol update as part of UpdateExport
UpdateExport DBUS does not handle protocol update. For example, when the
protocol is changes from NFSv3 to NFSv4 for an export, after UpdateExport
none of hte mounts worked.

This fix allows updating the protocols as part of UpdateExport. Updating to
NFSv4 was the problem because the new path was not added to pseudo FS

Change-Id: I6a78ec85fd2a48064161420d9459ecedebfad68f
Signed-off-by: Sriram Patil <sriramp@vmware.com>