On create export, refresh file system table
On each export created, we remove any unused (unclaimed) file systems
from the table and scan mtab for file systems required for the export:

- Any file system that leads to the export
- The file system containing the export root
- Any sub-mounted file systems

Otherwise, we exclude all other file systems to reduce scan time.

A stat() system call is made on the export root, and if it returns
EAGAIN we retry the stat() call to give autofs a chance to mount the
file system.

The release_posix_file_system() function has a parameter added to
indicate if it is ok to unclaim file systems. When removing file
systems on reload, we MUST NOT unclaim any file systems since there
may be references to them.

Since an FSAL_VFS unexport removes the unexported file systems on the
unexport of the LAST export using a file system, that should provide a
way to clean up file systems that have been force unmounted for some
reason.

NOTE: Unexport when nested exports are involved is fraught with issues.
      It MAY be true that unexporting an export that is within (under)
      another export would allow any file systems under the export that
      was removed can become available by file system crossing discovery
      during lookup. Removing an export that has exports within (under)
      it without having first removed those lower exports will probably
      not result in good things happening.

NOTE: FSAL_GPFS does not have some of the enhancements that FSAL_VFS
      does so the above may not be so true of FSAL_GPFS.

Change-Id: I6e94ef37300906a75f5ed2b5323ec0629fa51e34
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>