Update patch set 4
Patch Set 4: Code-Review-1

OK, there IS a reason we don't include the related owner in the hash and compare...

LOCKT provides a lock owner opaque without ANY connection to an open owner.

With this patch, LOCKT with same owner WILL conflict with actual lock from that lock owner. This is because to find the lock owner, a key is built from the lock owner provided. The key for LOCKT can not include the related owner, and thus does not match.

Please explain the situation where you have two lock owners with different related open owners.

One possible option is that we don't hash the related owner and on the compare, if one of the related owners is NULL then the compare succeeds. However, this would mean if you actually had two different lock owners with different open owners, LOCKT would match either one.

Anyway, this patch is not ready for inclusion.

Patch-set: 4
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1