Error => warning for Ganesha's FSAL_DIGEST_V4 request.
The mdcache has a hardcoded "I want a FSAL_DIGEST_V4 from you", which I
previously worked around with an innocuous ("V3 requests should be V3, V4 should
be V4"). But, on "boot", there is no active operation / NFS request that drives
the creation of the export, so the first one of these will fail. Not returning
to Ganesha the handle to our root is sure to cause problems.

So instead, since Ganesha doesn't look at it anyway, turn this into a
warning. If possible, I may try to make the internal handler detect that it's
being called for the root object and silently do that.

Change-Id: Ibdd8ee9e2af4b1275e7a0e07b7c4684708c4e2e0
Signed-off-by: Solomon Boulos <boulos@google.com>