Allow FSSTAT on any object.
I misread this originally, it's not just for the root of the filesystem. (This
came up in the final cthon basic test). With this in place, all the basic cthon
tests pass! But the readdir test is still super slow (10 minutes vs a handful of
seconds directly to an NFS server).

Change-Id: I5e7d7d8ed7a6a0ab678451548d1d027a0b5793d3
Signed-off-by: Solomon Boulos <boulos@google.com>