Don't update fs_info.maxfilesize based on the backend.
nlm_util (incorrectly) treats offsets > fs_info.maxfilesize as an opportunity to
exit early. This was introduced in c811fe9323 to try to avoid waiting for errors
downstream, except the cthon locking tests generate it and get the right answer
(NFSD will happily let you specify UINT64_MAX as an offset). dang@redhat.com
worked around this for the NFSv4 lock routines in 3d069bf, and we probably
should do the same. For now though, I don't think other FSALs update
fs_info.maxfilesize, so I'll just warn that we're not updating it.

Change-Id: Ibb494b5e7a90201df5bc0c30252ef3e495bf0f58
Signed-off-by: Solomon Boulos <boulos@google.com>