Update patch set 2
Patch Set 2:

(2 comments)

Please let me know, if its ok to fail grace in this case, or we should fix it in reaper?

if (!nfs_in_grace()) {
    if (atomic_fetch(&grace_status) & CHANGE_REQ)
        nfs_start_grace();
    else if (recovery_backend->maybe_start_grace)
        recovery_backend->maybe_start_grace()
}
This is not enough I have to pass the args to nfs_start_grace, will need another global variable to save that and pass here.
if multiple grace requests happen then it would be something like this.
1>nfs_get_grace_status()  -> Takes ref
2>nfs_start_grac() -> Sees ref, sets CHANGE_REQ, save "nfs_grace_start_t *gsp"
3>nfs_start_grac() -> Sees ref, sets CHANGE_REQ, save "nfs_grace_start_t *gsp", overwrite old.
4>nfs_put_grace_status() -> Checks CHANGE_REQ, starts reaper()

We could end up skipping one of event data in this way, if we overwrite in "3".
Other way could be to keep on adding these event data to the list and iterate through that list in reaper.

nfs_start_grace(gsp) {
     if (!was_grace && (old & GRACE_STATUS_COUNT_MASK)) {
         mutex_lock(gsp_list);
         list_add(gasp gsp_list);
         mutex_unlock(gsp_list);
         goto out
     }
}

nfs_maybe_start_grace() {
    if (!nfs_in_grace()) {
        if (atomic_fetch(&grace_status) & CHANGE_REQ)
            mutex_lock(gsp_list);
            for each gsp in gsp_list(nfs_start_grace(gsp));
            mutex_unlock(gsp_list);    
        else if (recovery_backend->maybe_start_grace)
            recovery_backend->maybe_start_grace()
    }
}

Patch-set: 2