Update patch set 2
Patch Set 2: -Code-Review

> nfs_get_grace_status()  -> Takes ref
> nfs_start_grac() -> Sees ref, sets CHANGE_REQ
> nfs_put_grace_status() -> Checks CHANGE_REQ, starts reaper()
> 
> Who will reset CHANGE_REQ?

Ok, got it. Yes, that is a problem.

I think the right approach is probably to just define a flag in nfs_grace_start_t that is only set by the dbus handler. If that flag is set, then have nfs_start_grace return an EAGAIN error or something, if we're not already in grace and there is an elevated refcount.

Then the dbus handler could either just keep polling after a short timeout, or you could add a new condition variable for it to wait on.

Patch-set: 2
Reviewer: Gerrit User 1008909 <1008909@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0