Make fullpath and pseudopath gsh_refstr and keep a ref in op_ctx
The pseudopath is going to be changeable in a following patch so we
must make a stable version easily accessible. Use gsh_refstr to
accomplish that.

Also, for consistency handle fullpath the same way. It may actually
be possible to allow changing of it also and in the meantime, code
that picks one or the other of pseudopath and fullpath will not have
to deal with different rules for access to those strings.

Change-Id: I40270796f6508f1d2f9cc7659347ab24189a03ff
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>