Make fullpath and pseudopath gsh_refstr and keep a ref in op_ctx
The pseudopath is going to be changeable in a following patch so we
must make a stable version easily accessible. Use gsh_refstr to
accomplish that.

Also, for consistency handle fullpath the same way. It may actually
be possible to allow changing of it also and in the meantime, code
that picks one or the other of pseudopath and fullpath will not have
to deal with different rules for access to those strings.

Since many of the references to fullpath and pseudopath are related to
op_ctx->ctx_export, take advantage of the standardized management of the
export in ctx_export to also take and release gsh_refstr references
to fullpath and pseudopath into ctx_fullpath and ctx_pseudopath, and if
an export is not present, make those references to the string
"No Export". This then allows most of the code that just wants to log
the string to just use the string from the op_ctx.

For those places that use an export other than ctx_export and need
access to the string, a struct tmp_export_paths with attendant
functions to manage the gsh_refstr references is provided.

The config processing is simplified by attaching the original string
as a simple string however these strings are ONLY available during
config processing and the entire use of the string is protected by
a lock.

A saved export also carries its gsh_refstr references to reduce the
need to reacquire a reference.

A junction also carries its own gsh_refstr reference to the pseudopath
to simplify access to that string.

Change-Id: I40270796f6508f1d2f9cc7659347ab24189a03ff
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>