In mdc_open2_by_name() after lookup check if obj_handle.type==REGULAR_FILE
In mdc_open2_by_name() after mdc_lookup() completes successfully
and prior to calling FSALs open2(..) we don't check the
entry->obj_handle.type to verify if the object is for a regular file.
It may be possible that prior to calling mdc_lookup() in
mdc_open2_by_name(), another thread working on a different client request
may have created a symbolic link with the same name. Using an object created
for a symbolic link in nfs4_op_open() code path lead to a crash.
Fixed this by adding a check in mdc_open2_by_name() to verify
entry->obj_handle.type is REGULAR_FILE and then proceed to call a FSALs
open2(..).

Change-Id: Ib86b53f1c7bb6652509ef98575f6b1e8ff2fde82
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>