Rename FSAL_PROXY => FSAL_PROXY_V4.
Now that there will be a PROXY_V3, Frank asked me to
rename this existing proxy to PROXY_V4. This change
just does the giv mv rename as well as fixing up
the CMake-scripts to point to the right thing.

Note that we are not changing the libfsalproxy.so library name/output,
because that would cause configurations looking for:

FSAL {
  NAME = PROXY
}

to fail due to the way load_fsal loads FSALs (i.e., it just takes the
name and maps it to an .so, then hands it to dlopen).

Change-Id: Id7eea79a77ca57c59e2d983243abdfed9070dd66
Signed-off-by: Solomon Boulos <boulos@google.com>