Rename the V4 proxy names from pxy_ to proxyv4.
As part of making the V3 proxy, ffilz asked that I rename a lot of the
code in the existing V4 proxy. One caveat is that load_fsal performs a
dlopen() of the name of the FSAL given in a config, so we cannot rename
the output PROXY_V4 unless we wish to break existing configs. Seems like
a future breaking change.

Change-Id: Ib7c94b0480cf9b05f303591c20585345e5e3e711
Signed-off-by: Solomon Boulos <boulos@google.com>