Update patch set 2
Patch Set 2:

I think this is not the correct solution.  It will break the case-insensitive nature of the rest of the config file, which I think we want to keep.  Probably we need a flag on save_token() to indicate if it should be case sensitive, and only set it to TRUE for strings (or maybe only double-quoted strings?)

Patch-set: 2
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>