Update patch set 2
Patch Set 2:

Please do some testing, but it looks like this is just managing the memory use for tokens to reduce duplication. With the change to strcmp from strcasecmp, all it means is each capitalization form of a token will be stored separately, but THIS str.*cmp is NOT what matches a token with how it's used in the syntax, there's a separate strcasecmp later to match for example EXPORT or Export or export to the syntax for an export block. This should at worst result in 3 stores of any given token that is used many times for the three most likely forms I just listed. Config files that are consistently capitalized will just store a single form of each token.

Patch-set: 2
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>