Breaking rename of libfsalproxy => libfsalproxy_v4.
dang@ suggested that since V4 is a while away, it would be better to rip
the bandaid now on telling people to update their configs. Keeping this
as a separate change hopefully makes it easier for people to patch the
other proxy changes into their backported tree if needed without forcing
a config update.

Change-Id: I08f6e4dc9355612e2560fd33fab1d23ae51eb1a0
Signed-off-by: Solomon Boulos <boulos@google.com>