FSAL_UP: add new "shrink" upcall
Ceph MDS's can request that the client shrink their caps cache in
response to memory pressure, and libcephfs now has a way to set
callbacks that get called when it wants to drop caps for an inode.
Add a new shrink upcall that FSAL_CEPH can call for this.

The upcall will find the entry and test its refcount. If the refcount is
2, then we know that there are no other references than the one we
currently hold and the sentinel reference. That should mean that it's
safe to drop it from the hash and try to kill it off.

Change-Id: Ib195569c88563235014fe1b7e6c37ea1333ceeea
Signed-off-by: Jeff Layton <jlayton@redhat.com>