Split state_lock into st_lock and jct_lock
The state_hdl state_lock is used for two different things.

When used for state (locks, shares, delegations) protection, it
is only used exclusively (i.e. write) so could be just a mutex.

When used for junction protection, it is used as a rwlock.

Also, it's handy for the locks to have a different order with
respect to the export lock. By breaking up and reversing the
order for jct_lock we eliminate a place where we drop the export
lock, take and drop the jct_lock and then retake the export lock.

And separating them makes the code a bit easier to read.

Also caught one place where the state_lock was not taken with
STATELOCK_lock.

Change-Id: I1ceb25d34c1ebae183fc142b2298a3468e6a1b02
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>