FSAL_CEPH: set ino_release_cb on mount
Newer versions of libcephfs allow the application to register
callbacks that will be called when the client has been requested
to shrink the number of caps it holds.

When libcephfs calls into the application to request releasing an
inode, try to satisfy that request by calling the try_release upcall.

Change-Id: Ic89f5f3aa1e1474c1aa3d4e59e9f69248266ab18
Signed-off-by: Jeff Layton <jlayton@redhat.com>