Update patch set 1
Patch Set 1: Code-Review-1

I don't think this is valid.  If we reload the chunk, we can't be sure that the exact same set of dirents are still in the same order as last time, and so the next_ck of the reloaded chunk may not be the same as it was the first time.  We can only set next_ck if we've actually found the first ck of the next chunk.

Patch-set: 1
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1