Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> I'm curious why you didn't re-use ERR_FSAL_DELAY for "Yes, this exists, but I'm busy over here"

There's a subtle difference.

ERR_FSAL_DELAY is really analogous to NFS4ERR_DELAY, which is really more similar to POSIX EAGAIN. That implies that waiting a bit and trying again might help.

This is more like EBUSY. The thing you want is in use, and it may be for the forseeable future. You're welcome to try again, but you shouldn't have any expectation that it might help.

Patch-set: 1