Add struct saved_export_context to standardize saving op_ctx bits
There are times when a process just wants to make a temporary export
change in the op_context but doesn't want to craft a whole new
op_context. This addition presents a standard way of saving the
export related bits of an op_context in order to be able to restore it
after the temporary change.

Change-Id: Ifad0c091c10687e75e7a2c39eabaf9a5b7d61986
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>