Update patch set 8
Patch Set 8:

Okay, so this should work, but can you explain what the advantage is?  What you're trying to save?  It should work correctly without this, so is it an optimization?  If so, how much difference does it make?

(And, so I don't forget, I have some nits if we end up taking this)

Patch-set: 8
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>