stats: Adds json output to some stats command in ganesha_stats.py
This adds the support of the json format for most commands by adding
a new method to all report objects.

This is done in two step, one method is generating a dict and the
other simply dump it in the json format. In the future the text format
could also leverage the report method, but it was kept separate to kept
changes at their minimal.

Signed-off-by: Romain Vaillant <romain.vaillant@scality.com>
Change-Id: I849b751c6735847e8152713bf2807c656dac0ea3