Update patch set 3
Patch Set 3:

This is strange, because I run every build with ASAN, and it works fine for me.  Before we go messing with the linking, I'd like to know why it works for me and fails for you.

Patch-set: 3
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>