Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > How about just taking the attr_lock for read around the write() call?  That would block mdcache_refresh_attrs() (which needs it taken for write).
> 
> That would also block write()s from happening during the mdcache_refresh_attrs(). Also, these write()s would anyways need to call mdcache_refresh_attrs() after they're done, so we haven't saved any getattr calls, just blocked write()s, no?

The goal isn't to save calls to refresh_attrs, but just to make sure that attrs are valid when we think they're valid.  Blocking writes when we're refreshing could be a problem; depending on how often these collide.

I'm a bit leery of adding a new mechanism to track validity to an already complex setup.  The *best* solution would be to have write() return valid attributes we can save to the cache. But I'm not sure that's possible.

Patch-set: 1