Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > How about just taking the attr_lock for read around the write() call?  That would block mdcache_refresh_attrs() (which needs it taken for write).
> > 
> > That would also block write()s from happening during the mdcache_refresh_attrs(). Also, these write()s would anyways need to call mdcache_refresh_attrs() after they're done, so we haven't saved any getattr calls, just blocked write()s, no?
> 
> The goal isn't to save calls to refresh_attrs, but just to make sure that attrs are valid when we think they're valid.
I meant that the goal of the cache is to save refresh_attrs calls, and using a lock won't save any more than using the generation mechanism we've proposed. 

> Blocking writes when we're refreshing could be a problem; depending on how often these collide.
Yes the chance of that happening probably depends on the latency and parallelism of the underlying filesystem, but at least on our runs of https://github.com/phdeniel/ltpfs/blob/master/testcases/kernel/fs/ftest/ftest03.c over wekafs it happens quite a lot, and in the writes in which the file's size needs to grow, it eventually causes data corruption.

> I'm a bit leery of adding a new mechanism to track validity to an already complex setup.
I understand the need for caution, and the mechanism not being fully intuitive, but do you see currently any potential of it breaking other things? i.e. AFAIK it should only cause invalidation of the cache in some cases, which in itself should be a safe thing to do, no?

> The *best* solution would be to have write() return valid attributes we can save to the cache. But I'm not sure that's possible.
How could that be possible with the FSAL_VFS, which uses posix syscalls?

Patch-set: 1