Add functions for encoding and decoding POSIX ACLs and the GETACL and SETACL RPCs.
The actual access decisions are performed using the ACCESS RPC which is part of NFSv3 proper, and is independent of acls.
The GETACL and SETACL RPCs are mainly used by tools like getfacl and setfacl.In addition, for files created inside directories that have a default acl,SETACL is used at file create time to implement the POSIX ACL file create semantics.

Change-Id: I5b73caf733c54003370abe81200f9ac5cc8b6abc
Signed-off-by: Luo Muyao <luo.muyao@zte.com.cn>
Signed-off-by: Shangzhong Zhu <zhu.shangzhong@zte.com.cn>