Update patch set 1
Patch Set 1:

It looks like this is converting over the wire NFSACL ACLs (effectively POSIX ACLs) to FSAL ACLs which are NFSv4 ACLs. Since POSIX ACL <-> NSF4 ACL conversion is not perfect, and both FSAL_GLUSTER and FSAL_VFS actually implement POSIX ACLs on the file system, I think we should add a POSIX ACL type to the FSAL and have struct attrlist indicate if the ACL is POSIX or NFS4 so we don't have to do unnecessary lossy ACL conversions.

Patch-set: 1